# docs/web-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ajib API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/auth/challenge": {
      "post": {
        "summary": "Challenge",
        "operationId": "challenge_api_v1_auth_challenge_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChallengeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/consume": {
      "post": {
        "summary": "Consume",
        "operationId": "consume_api_v1_auth_consume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConsumeInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/me/language": {
      "put": {
        "summary": "Language",
        "operationId": "language_api_v1_me_language_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LanguageInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/orders": {
      "post": {
        "summary": "Create Order",
        "operationId": "create_order_api_v1_orders_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckoutInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/payments/{payment_id}/receipt": {
      "post": {
        "summary": "Receipt",
        "operationId": "receipt_api_v1_payments__payment_id__receipt_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_receipt_api_v1_payments__payment_id__receipt_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/receipts/{receipt_id}": {
      "get": {
        "summary": "View Receipt",
        "operationId": "view_receipt_api_v1_receipts__receipt_id__get",
        "parameters": [
          {
            "name": "receipt_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Receipt Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/referrals/wallet": {
      "put": {
        "summary": "Referral Wallet",
        "operationId": "referral_wallet_api_v1_referrals_wallet_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WalletInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/referrals/attribution": {
      "post": {
        "summary": "Referral Attribution",
        "operationId": "referral_attribution_api_v1_referrals_attribution_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttributionInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/payments/{payment_id}/review": {
      "post": {
        "summary": "Review",
        "operationId": "review_api_v1_admin_payments__payment_id__review_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Payment Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttributionInput": {
        "properties": {
          "code": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Code"
          }
        },
        "type": "object",
        "required": [
          "code"
        ],
        "title": "AttributionInput"
      },
      "Body_receipt_api_v1_payments__payment_id__receipt_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_receipt_api_v1_payments__payment_id__receipt_post"
      },
      "ChallengeInput": {
        "properties": {
          "storefront": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60,
                "pattern": "^[a-z0-9-]+$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Storefront"
          }
        },
        "type": "object",
        "title": "ChallengeInput"
      },
      "CheckoutInput": {
        "properties": {
          "plan_id": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Plan Id"
          },
          "method": {
            "type": "string",
            "enum": [
              "card",
              "crypto"
            ],
            "title": "Method"
          },
          "username": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "Username"
          },
          "server_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Server Id"
          },
          "reserved": {
            "type": "boolean",
            "title": "Reserved",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "plan_id",
          "method"
        ],
        "title": "CheckoutInput"
      },
      "ConsumeInput": {
        "properties": {
          "challenge": {
            "type": "string",
            "maxLength": 64,
            "minLength": 20,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Challenge"
          }
        },
        "type": "object",
        "required": [
          "challenge"
        ],
        "title": "ConsumeInput"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LanguageInput": {
        "properties": {
          "language": {
            "type": "string",
            "enum": [
              "fa",
              "en",
              "tk",
              "ru"
            ],
            "title": "Language"
          }
        },
        "type": "object",
        "required": [
          "language"
        ],
        "title": "LanguageInput"
      },
...
      "ReviewInput": {
        "properties": {
          "approve": {
            "type": "boolean",
            "title": "Approve"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "minLength": 3,
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "approve",
          "reason"
        ],
        "title": "ReviewInput"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WalletInput": {
        "properties": {
          "address": {
            "type": "string",
            "maxLength": 256,
            "minLength": 10,
            "title": "Address"
          }
        },
        "type": "object",
        "required": [
          "address"
        ],
        "title": "WalletInput"
      }
    }
  }
}